Moved s3 to separate remotes module
